Modified OcclusionCuller, no longer shows occluded voxels
FOV seems to be slightly smaller than before and sometimes does not show points on the bottom of the plane.

FOV may be offset with a larger FOV parameters. Bottom of the plane may be occluded by steep angle and large voxel sizes. Would have to look up to see them properly